Clean-up lessons data
