+++ libraries/media/api/src/main/kotlin/ss/libraries/media/api/MediaRepository.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022. Adventech <info@adventech.io>
+ * Copyright (c) 2025. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -13,38 +13,23 @@
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
 
-package com.cryart.sabbathschool.test.di.repository
+package ss.libraries.media.api
 
-import app.ss.lessons.data.repository.media.MediaRepository
 import app.ss.models.media.AudioFile
 import app.ss.models.media.SSAudio
 import app.ss.models.media.SSVideosInfo
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flowOf
-import javax.inject.Inject
 
-class FakeMediaRepository @Inject constructor() : MediaRepository {
-
-    override fun getAudio(
-        lessonIndex: String
-    ): Flow<List<SSAudio>> = flowOf(emptyList())
-
-    override suspend fun findAudioFile(
-        id: String
-    ): AudioFile? = null
-
-    override suspend fun getPlayList(
-        lessonIndex: String
-    ): List<AudioFile> = emptyList()
-
-    override fun getVideo(
-        lessonIndex: String
-    ): Flow<List<SSVideosInfo>> = flowOf(emptyList())
+interface MediaRepository {
+    fun getAudio(lessonIndex: String): Flow<List<SSAudio>>
+    suspend fun findAudioFile(id: String): AudioFile?
+    suspend fun getPlayList(lessonIndex: String): List<AudioFile>
+    fun getVideo(lessonIndex: String): Flow<List<SSVideosInfo>>
 }
