+++ libraries/test_utils/build.gradle.kts
@@ -35,7 +35,6 @@ android {
 dependencies {
     implementation(projects.common.auth)
     implementation(projects.common.core)
-    implementation(projects.common.lessonsData)
     implementation(projects.libraries.lessons.model)
 
     implementation(libs.google.hilt.android)
