+++ libraries/test_utils/src/main/java/com/cryart/sabbathschool/test/di/FakeRepositoryModule.kt
@@ -24,10 +24,7 @@ package com.cryart.sabbathschool.test.di
 
 import app.ss.auth.AuthRepository
 import app.ss.auth.di.AuthBindings
-import app.ss.lessons.data.di.RepositoryModule
-import app.ss.lessons.data.repository.media.MediaRepository
 import com.cryart.sabbathschool.test.di.repository.FakeAuthRepository
-import com.cryart.sabbathschool.test.di.repository.FakeMediaRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.components.SingletonComponent
@@ -37,14 +34,10 @@ import dagger.hilt.testing.TestInstallIn
 @TestInstallIn(
     components = [SingletonComponent::class],
     replaces = [
-        RepositoryModule::class,
         AuthBindings::class
     ]
 )
 abstract class FakeRepositoryModule {
-
-    @Binds
-    abstract fun bindMediaRepository(impl: FakeMediaRepository): MediaRepository
 
     @Binds
     abstract fun bindAuthRepository(impl: FakeAuthRepository): AuthRepository
