+++ services/media/impl/src/main/kotlin/ss/services/media/impl/di/BindingsModule.kt
@@ -26,7 +26,9 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ActivityComponent
+import ss.libraries.media.api.MediaRepository
 import ss.libraries.media.api.SSMediaPlayer
+import ss.services.media.impl.MediaRepositoryImpl
 import ss.services.media.impl.SSMediaPlayerImpl
 
 @Module
@@ -35,4 +37,7 @@ abstract class BindingsModule {
 
     @Binds
     internal abstract fun bindSSMediaPlayer(impl: SSMediaPlayerImpl): SSMediaPlayer
+
+    @Binds
+    internal abstract fun bindMediaRepository(impl: MediaRepositoryImpl): MediaRepository
 }
